fix: uuid types in data structures
the current implementation assumes xandra always returns strings and
avoids loading uuids itself.

this usually works, but has a few drawbacks:
- we're forced to use embedded loaders for data structures, which should
  not be needed as the structures themselves declare their loaders
- these embedded loaders perform the UUID load themselves, attempting at
  loading a string UUID and resulting in an exception

By requiring xandra to send us UUIDs in binary format and performing the
loading ourselves, we avoid the issue above.

The drawback of this is that it is no longer possible to request UUIDs
in binary format to Xandra, but this did not work for values nested
inside maps/sets even before this refactor